docs: the typescript docs has been created
